Add assertion to check that the To field contains only one email and refactor email related functions to its own file
